+++ ui/constants/icons.js
@@ -100,3 +100,4 @@ export const SLIDERS = 'Sliders';
 export const SCIENCE = 'Science';
 export const ANALYTICS = 'BarChart2';
 export const PURCHASED = 'Key';
+export const CIRCLE = 'Circle';